app/requirements.txt
fastapi==0.110.0
paho-mqtt==1.6.1
psycopg2-binary==2.9.9
pytz==2023.3
typing_extensions==4.10.0
uvicorn==0.29.0
